squash migrations from development
