+++ src/backend/alembic/versions/1cbb45f9fe25_.py
@@ -1,0 +1,40 @@
+"""empty message
+
+Revision ID: 1cbb45f9fe25
+Revises: f5819b10ef2a
+Create Date: 2024-06-07 20:39:32.645630
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '1cbb45f9fe25'
+down_revision: Union[str, None] = 'f5819b10ef2a'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.alter_column('agents', 'version',
+               existing_type=sa.INTEGER(),
+               nullable=True)
+    op.alter_column('agents', 'temperature',
+               existing_type=sa.DOUBLE_PRECISION(precision=53),
+               nullable=True)
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.alter_column('agents', 'temperature',
+               existing_type=sa.DOUBLE_PRECISION(precision=53),
+               nullable=False)
+    op.alter_column('agents', 'version',
+               existing_type=sa.INTEGER(),
+               nullable=False)
+    # ### end Alembic commands ###
