--- src/backend/alembic/versions/bed33ee3a3dd_.py
@@ -1,32 +1,0 @@
-"""empty message
-
-Revision ID: bed33ee3a3dd
-Revises: f5819b10ef2a
-Create Date: 2024-06-07 19:16:49.278403
-
-"""
-
-from typing import Sequence, Union
-
-import sqlalchemy as sa
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision: str = "bed33ee3a3dd"
-down_revision: Union[str, None] = "f5819b10ef2a"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column("agents", "description", existing_type=sa.TEXT(), nullable=False)
-    op.alter_column("agents", "preamble", existing_type=sa.TEXT(), nullable=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column("agents", "preamble", existing_type=sa.TEXT(), nullable=True)
-    op.alter_column("agents", "description", existing_type=sa.TEXT(), nullable=True)
-    # ### end Alembic commands ###
