--- src/backend/alembic/versions/f5b13c574d22_.py
@@ -1,46 +1,0 @@
-"""empty message
-
-Revision ID: f5b13c574d22
-Revises: bed33ee3a3dd
-Create Date: 2024-06-07 19:29:57.960278
-
-"""
-
-from typing import Sequence, Union
-
-import sqlalchemy as sa
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision: str = "f5b13c574d22"
-down_revision: Union[str, None] = "bed33ee3a3dd"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column("agents", "version", existing_type=sa.INTEGER(), nullable=True)
-    op.alter_column("agents", "description", existing_type=sa.TEXT(), nullable=True)
-    op.alter_column("agents", "preamble", existing_type=sa.TEXT(), nullable=True)
-    op.alter_column(
-        "agents",
-        "temperature",
-        existing_type=sa.DOUBLE_PRECISION(precision=53),
-        nullable=True,
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column(
-        "agents",
-        "temperature",
-        existing_type=sa.DOUBLE_PRECISION(precision=53),
-        nullable=False,
-    )
-    op.alter_column("agents", "preamble", existing_type=sa.TEXT(), nullable=False)
-    op.alter_column("agents", "description", existing_type=sa.TEXT(), nullable=False)
-    op.alter_column("agents", "version", existing_type=sa.INTEGER(), nullable=False)
-    # ### end Alembic commands ###
